% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_function.R
\name{mid_rescaler}
\alias{mid_rescaler}
\title{internal version of mid_rescaler from ggplot2 package}
\usage{
mid_rescaler(mid, transform)
}
\arguments{
\item{mid}{The midpoint (in data value) of the diverging scale.\if{html}{\out{<br>}}Defaults to 0.}

\item{transform}{For continuous scales, the name of a transformation object or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and "time".\if{html}{\out{<br>}}A transformation object bundles together a transform, its inverse, and methods for generating breaks and labels. Transformation objects are defined in the scales package, and are called ⁠\verb{transform_<name>}⁠. If transformations require arguments, you can call them from the scales package, e.g \code{\link[scales]{transform_boxcox}}... You can create your own transformation with \code{\link[scales]{new_transform}}.}
}
\description{
internal version of mid_rescaler from ggplot2 package
}
\keyword{internal}
