% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pal_tableau}
\alias{pal_tableau}
\title{Palettes from Tableau}
\format{
A list with 2 levels
The first level is one of "regular", "sequential" and "diverging".
The seconde level is the name of the palette.
Note that some palettes in "sequential" starts with "map" is from CARTO.
}
\source{
\url{https://help.tableau.com/current/pro/desktop/en-us/viewparts_marks_markproperties_color.htm}\cr
\url{https://carto.com/carto-colors/}
}
\usage{
pal_tableau
}
\description{
Palettes from Tableau, could easily be called by the \code{$} and auto-complete feature of the R IDEs.
}
\keyword{datasets}
