% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_function.R
\name{rescale_gradient}
\alias{rescale_gradient}
\title{Rescale function for gradient}
\usage{
rescale_gradient(palette, direction, start, end)
}
\arguments{
\item{palette}{Palette from "paltableau" package. \if{html}{\out{<br>}} Try to type "pal_tableau$" and wait for hint of "regular", "sequential" or "diverging",\if{html}{\out{<br>}} select one, and type "$" to choose the ultimate palette.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{start}{The (corrected) color range in [0,1]⁠ at which the color map starts.}

\item{end}{The (corrected) color range in [0,1]⁠ at which the color map ends.}
}
\description{
Rescale function for gradient
}
\keyword{internal}
