% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{scale_color_tableau_binned_gradient}
\alias{scale_color_tableau_binned_gradient}
\title{Tableau color scales (binned)}
\usage{
scale_color_tableau_binned_gradient(
  palette = pal_tableau$sequential$blue_teal,
  na.value = "grey50",
  direction = 1,
  start = 0,
  end = 1,
  ...
)
}
\arguments{
\item{palette}{Palette from "paltableau" package. \if{html}{\out{<br>}} Try to type "pal_tableau$" and wait for hint of "regular", "sequential" or "diverging",\if{html}{\out{<br>}} select one, and type "$" to choose the ultimate palette.}

\item{na.value}{Colour to use for missing values}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{start}{The (corrected) color range in [0,1]⁠ at which the color map starts.}

\item{end}{The (corrected) color range in [0,1]⁠ at which the color map ends.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{binned_scale}}.}
}
\description{
Tableau color scales (binned)
}
\examples{
ggplot(iris)+
  geom_point(aes(Sepal.Length,Sepal.Width,color = Sepal.Length))+
  scale_color_tableau_binned_gradient(palette = pal_tableau$diverging$temperature,
                                      direction = 1,
                                      n.breaks = 7,
                                      start = 0.2,end = 0.8)
}
