% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{scale_color_tableau_discrete}
\alias{scale_color_tableau_discrete}
\title{Tableau color scales (discrete)}
\usage{
scale_color_tableau_discrete(
  palette = pal_tableau$regular$`Tableau 10`,
  direction = 1,
  na.value = "grey50",
  start = 0,
  end = 1,
  ...
)
}
\arguments{
\item{palette}{Palette from "paltableau" package. \if{html}{\out{<br>}} Try to type "pal_tableau$" and wait for hint of "regular", "sequential" or "diverging",\if{html}{\out{<br>}} select one, and type "$" to choose the ultimate palette.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{na.value}{Colour to use for missing values}

\item{start}{The (corrected) color range in [0,1]⁠ at which the color map starts.}

\item{end}{The (corrected) color range in [0,1]⁠ at which the color map ends.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}}.}
}
\description{
Tableau color scales (discrete)
}
\examples{
ggplot(iris)+
 geom_col(aes(Sepal.Length,Sepal.Width,fill = Species))+
 scale_fill_tableau_discrete(palette = pal_tableau$diverging$orange_blue,
                              direction = -1,
                              start = 0.2,end = 0.8)
}
