% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_tableau_gradient.R
\name{scale_fill_tableau_gradient}
\alias{scale_fill_tableau_gradient}
\title{Tableau fill scales (gradient)}
\usage{
scale_fill_tableau_gradient(
  palette = pal_tableau$sequential$blue_teal,
  direction = 1,
  na.value = "grey50",
  start = 0,
  end = 1,
  guide = "colourbar",
  ...
)
}
\arguments{
\item{palette}{Palette from "paltableau" package. \if{html}{\out{<br>}} Try to type "pal_tableau$" and wait for hint of "regular", "sequential" or "diverging",\if{html}{\out{<br>}} select one, and type "$" to choose the ultimate palette.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{na.value}{Colour to use for missing values}

\item{start}{The (corrected) color range in [0,1]⁠ at which the color map starts.}

\item{end}{The (corrected) color range in [0,1]⁠ at which the color map ends.}

\item{guide}{Type of legend. Use 'colourbar' for continuous colour bar, or 'legend' for discrete colour legend.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{continuous_scale}}.}
}
\description{
Tableau fill scales (gradient)
}
\examples{
library(ggplot2)
library(paltableau)
ggplot(iris)+
  geom_col(aes(Sepal.Length, Sepal.Width, fill = Sepal.Length))+
  scale_fill_tableau_gradient(palette = pal_tableau$diverging$temperature,
                              direction = 1,
                              start = 0.2,end = 0.8)
}
