% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_tableau_steps.R
\name{scale_fill_tableau_steps}
\alias{scale_fill_tableau_steps}
\title{Binned gradient fill scales with Tableau palettes}
\usage{
scale_fill_tableau_steps(
  palette = pal_tableau$sequential$blue_teal,
  na.value = "grey50",
  direction = 1,
  start = 0,
  end = 1,
  breaks = NULL,
  guide = "coloursteps",
  ...
)
}
\arguments{
\item{palette}{Palette from "paltableau" package. \if{html}{\out{<br>}} Try to type "pal_tableau$" and wait for hint of "regular", "sequential" or "diverging",\if{html}{\out{<br>}} select one, and type "$" to choose the ultimate palette.}

\item{na.value}{Colour to use for missing values}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{start}{The (corrected) color range in [0,1]⁠ at which the color map starts.}

\item{end}{The (corrected) color range in [0,1]⁠ at which the color map ends}

\item{breaks}{The breaks for bins}

\item{guide}{Type of legend. Use guide names like 'coloursteps' or 'bins', or use     a function used to create a guide. See \code{\link[ggplot2]{guides}} for more information.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{binned_scale}}.}
}
\description{
Binned gradient fill scales with Tableau palettes
}
\examples{
library(ggplot2)
library(paltableau)
ggplot(iris)+
  geom_col(aes(Sepal.Length, Sepal.Width, fill = Sepal.Length))+
  scale_fill_tableau_steps(palette = pal_tableau$diverging$temperature,
                                      direction = 1,
                                      n.breaks = 7,
                                      guide = guide_bins(theme = theme(legend.ticks = element_blank(),legend.axis.line = element_line(color = "transparent"))),
                                      start = 0.2,end = 0.8)
}
